MDEV-9312: storage engine not enforced during galera cluster replication
enforce_storage_engine was not initialized for wsrep applier
threads, as a result it was not honored on slave nodes.

The commit deals with this issue in part by fixing the
condition that wrongly disallowed enforced plugin to be
enabled for wsrep threads. This, however, is not sufficient
for startup threads, as they might get initialized even before
plugins. Fixed by re-initializing wsrep threads plugin
variables only after their global counterparts are initialized.

Added a test case.