Restore parallel replication async deadlock kill
Restore the original code for parallel replication deadlock kill.
When a deadlock kill is detected inside the storage engine, the kill
is not done immediately, to avoid calling back into the storage engine
kill_query method with various lock subsystem mutexes held. Instead the
kill is queued and done later by a slave background thread.

This avoids the need for locking hacks in kill_query, and for the
TokuDB implementation solves a problem with the fine-grained
per-locktree locking order.

With this patch, I was not able so far to reproduce neither the hangs
nor the crash/assertion seen previously, though more tests will be
meeded to say this conclusively.