Also report waits in case a deadlock is detected.
If transaction T1 tries to get a lock held by a later transaction T2,
and T1 fails due to deadlock, then we can be pretty sure that T1 will
get a row conflict with T2 again. So we might as well get T2 deadlock
killed as soon as possible to allow parallel replication to continue.