Fix interaction between threads during parallel replication deadlock kill
With asynchronous deadlock kill, we need to make sure that the to-be-killed
thread does not go away while the kill is pending in the queue of the
background thread.

Do this by marking the thread in thd->rgi_slave->killed_for_retry that a
kill is pending, and having threads wait for that flag to clear.